+++ robot_descriptions/_descriptions.py
@@ -115,6 +115,7 @@ DESCRIPTIONS: Dict[str, Description] = {
     "rhea_description": Description(Format.URDF),
     "robotiq_2f85_description": Description(Format.URDF),
     "robotiq_2f85_mj_description": Description(Format.MJCF),
+    "robotiq_2f85_v4_mj_description": Description(Format.MJCF),
     "romeo_description": Description(Format.URDF),
     "sawyer_mj_description": Description(Format.MJCF),
     "shadow_hand_mj_description": Description(Format.MJCF),