+++ examples/load_in_mujoco.py
@@ -7,7 +7,7 @@
 """
 Show a robot description, specified from the command line, in MuJoCo.
 
-This example requires MuJoCo, which is installed by ``pip install mujoco``.
+This example requires MuJoCo, which is installed by `pip install mujoco`.
 """
 
 import argparse