+++ examples/show_in_mujoco.py
@@ -7,8 +7,9 @@
 """
 Show a robot description, specified from the command line, using MuJoCo.
 
-This example requires MuJoCo, which is installed by `pip install mujoco`, and
-the MuJoCo viewer installed by `pip install mujoco-python-viewer`.
+This example is equivalent to `python -m robot_descriptions show_in_mujoco`. It
+requires MuJoCo, which is installed by `pip install mujoco`, and the MuJoCo
+viewer installed by `pip install mujoco-python-viewer`.
 """
 
 import argparse