+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-### Added 
+### Added
 
-- Description: Robotiq 2F-85 (MJCF V4)
+- Description: Robotiq 2F-85 (MJCF V4) (thanks to @peterdavidfagan)
 
 ## [1.14.0] - 2025-01-07