+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Description: AgileX PiPER (MJCF)
 - Description: AgileX PiPER (URDF)
+- Description: Robot Soccer Kit
 - Description: Robotiq 2F-85 (MJCF V4) (thanks to @peterdavidfagan)
 
 ### Fixed