+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-### Fixed
+### Added
 
-- CICD: Correct conda-forge dependencies in Pinocchio loader test environment
+- CICD: Switch to conda-forge dependencies in Pinocchio loader test environment
 
 ## [1.15.0] - 2025-03-06