+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - CICD: Isolate loader tests in testing submodules
+- CICD: Remove version pin on NumPy < 2
 
 ## [1.15.0] - 2025-03-06